% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestedAnalysis-package.R
\docType{package}
\name{forestedAnalysis-package}
\alias{forestedAnalysis}
\alias{forestedAnalysis-package}
\title{forestedAnalysis: Spatial Cross-Validation and AOA Analysis of Forest Cover}
\description{
A research compendium analyzing forest cover data in Washington and Georgia. It evaluates the Area of Applicability (AOA) and demonstrates model failure during spatial extrapolation.
}
\author{
\strong{Maintainer}: Rob Wiederstein \email{khuon68@gmail.com}

}
\keyword{internal}
