% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_state_topo}
\alias{plot_state_topo}
\title{Create Single State Topo Plot}
\usage{
plot_state_topo(data, boundary_sf, raster_path, state_name)
}
\arguments{
\item{data}{Dataframe containing state data.}

\item{boundary_sf}{sf object for the state boundary.}

\item{raster_path}{Path to the elevation .tif file.}

\item{state_name}{String name of the state.}
}
\description{
Generates a ggplot object for a single state's topography.
}
