% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{theme_forestry_spatial}
\alias{theme_forestry_spatial}
\title{Standardized Spatial Theme}
\usage{
theme_forestry_spatial()
}
\description{
High-visibility map theme for presentations.
}
\examples{
\dontrun{
# How to apply the theme to a basic sf plot:
library(ggplot2)
library(sf)
nc <- st_read(system.file("shape/nc.shp", package="sf"))
ggplot(nc) + 
  geom_sf() + 
  theme_forestry_spatial()
}
}
