% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forested-package.R
\docType{package}
\name{forested-package}
\alias{forested-package}
\title{forested: Forest Attributes in U.S. States}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A small subset of plots throughout the U.S. are sampled and assessed "on-the-ground" as forested or non-forested by the U.S. Department of Agriculture, Forest Service, Forest Inventory and Analysis (FIA) Program, but the FIA also has access to remotely sensed data for all land in the country. The 'forested' package contains data frames intended for use in predictive modeling applications where the more easily-accessible remotely sensed data can be used to predict whether a plot is forested or non-forested. Currently, the package provides data for Washington and Georgia.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/forested}
  \item \url{https://simonpcouch.github.io/forested/}
  \item Report bugs at \url{https://github.com/simonpcouch/forested/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

Authors:
\itemize{
  \item Grayson White \email{graysonwhite13@gmail.com} (\href{https://orcid.org/0000-0003-4993-2792}{ORCID})
  \item Hannah Frick \email{hannah@posit.co} (\href{https://orcid.org/0000-0002-6049-5258}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
}

}
\keyword{internal}
